+++ packages/horseman-components/src/__tests__/index-test.js
@@ -10,10 +10,14 @@ const components = [
   "Container",
 ];
 
+const exportedUtils = ["media"];
+
 describe("horseman.js index", () => {
   describe("should export", () => {
     test("the Correct Number of items", () => {
-      expect(Object.keys(HorsemanComponents)).toHaveLength(components.length);
+      expect(Object.keys(HorsemanComponents)).toHaveLength(
+        components.length + exportedUtils.length,
+      );
     });
     components.forEach(component => {
       test(component, () => {
@@ -21,6 +25,15 @@ describe("horseman.js index", () => {
         expect(HorsemanComponents[component]).toEqual(
           // eslint-disable-next-line import/no-dynamic-require
           require(`../components/${component}`).default,
+        );
+      });
+    });
+
+    exportedUtils.forEach(e => {
+      test(e, () => {
+        expect(HorsemanComponents.utils[e]).toEqual(
+          // eslint-disable-next-line import/no-dynamic-require
+          require(`../styleUtils/${e}`).default,
         );
       });
     });