+++ packages/horseman-components/src/components/Picture/Picture.jsx
@@ -6,18 +6,18 @@ import { imageType } from "../../types";
 const StyledPicture = styled.picture`
   img {
     max-width: 100%;
+    width: 100%;
     height: auto;
     display: block;
   }
 `;
 
 const Picture = ({ src, srcset, alt, ...rest }) => {
-  const items = srcset.map(sourceSet => (
-    <source
-      key={sourceSet.src}
-      media={`(min-width:${sourceSet.width}px)`}
-      srcSet={sourceSet.src}
-    />
+  // Sort the srcset by width from largest to smallest
+  const sortedSrcset = srcset.slice().sort((a, b) => a.width < b.width);
+
+  const items = sortedSrcset.map(s => (
+    <source key={s.src} media={`(min-width:${s.width}px)`} srcSet={s.src} />
   ));
 
   return (